Several modifications on BioSystem, and BioProcessor (but no API changes!)
By now, the detailed sensitivity analysis should work as discussed.  In
particular, there is no need to compute the overview analysis anymore 
(hopefully...).  Parameter identification, however, might still be dodgy, 
we will see...

Additionally, there is a field 'itmax' in IOpt, to have a common max.
iteration number that is set to the Wk-s accordingly.